+++ CHANGELOG.md
@@ -36,7 +36,8 @@ To Document:
 
 TO DO:
 * IN PROGRESS: Factor out RunResult
-  * NEXT: Convert beforeAll/afterAll from TestCaseResult[] to RunResult[]
+  * NEXT: Convert TestRenderer to use .beforeEach/.afterEach instead of .beforeEach_OLD/.afterEach_OLD
+  * Convert beforeAll/afterAll from TestCaseResult[] to RunResult[]
   * Convert TestResult.suite() to TestSuiteResult.create() and TestResult.testCase() to TestCaseResult.create()
   * Revise TestSuiteResult.create() to match signature of other result factories
   * THEN: Modify TestSuite to return RunResult when running test function rather than TestCaseResult